requirements.txt
django
djangorestframework
python-decouple
psycopg2-binary
drf-yasg
gunicorn
dj-database-url
whitenoise
django-filter
django-rest-swagger
openapi-codec
djangorestframework-simplejwt
urllib3==1.26.6
requests